python/requirements.txt
baidu-aip==4.16.13
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
idna==3.6
playsound==1.3.0
requests==2.31.0
urllib3==2.1.0
